chore: move elf_cache down into runtime
