+++ engine/src/flutter/runtime/shorebird/BUILD.gn
@@ -1,0 +1,15 @@
+import("//flutter/common/config.gni")
+
+source_set("elf_cache") {
+  sources = [
+    "elf_cache.cc",
+    "elf_cache.h",
+    "elf_mapping.cc",
+    "elf_mapping.h",
+  ]
+
+  deps = [
+    "//flutter/fml",
+    "//flutter/runtime:libdart",
+  ]
+}
