+++ engine/src/flutter/shell/common/shorebird/elf_cache_unittests.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "flutter/shell/common/shorebird/elf_cache.h"
+#include "flutter/runtime/shorebird/elf_cache.h"
 
 #include <string>
 #include <vector>
